+++ src/components/Game/GameWinners/GameWinners.vue
@@ -58,7 +58,14 @@ import GameResult from "@/components/shared/Game/GameResult/GameResult";
 export default {
     name: "GameWinners",
     components: { GameResult, GameSummaryButton, GameReviewButton, GameReviewModal, GameSummaryModal, PlayerCard },
-    computed: { ...mapGetters("game", { game: "game" }) },
+    computed: {
+        ...mapGetters("game", { game: "game" }),
+        ...mapGetters("audioManager", { audioManager: "audioManager" }),
+    },
+    created() {
+        this.audioManager.stopNightMusic();
+        this.audioManager.stopDayMusic();
+    },
     methods: {
         confirmRestartGame() {
             return Swal.fire({