+++ src/components/shared/Game/Role/RoleText.vue
@@ -31,7 +31,7 @@ export default {
             } else if (groups.includes(this.role) || this.role === "sheriff") {
                 prefix = "the";
             } else {
-                prefix = this.game.getPlayersWithRole(this.role).length === 1 ? "the" : "a";
+                prefix = this.game.getPlayersWithRole(this.role).length > 1 ? "a" : "the";
             }
             return capitalize(this.$t(`Role.${prefix}.${this.role}`));
         },